Fix about menu
