+++ electron/menu/common.js
@@ -1,30 +1,3 @@
-import { app, dialog, BrowserWindow } from 'electron';
-import multiline from 'multiline-template';
-
-const remote = null; // TODO
-
-console.log(app)
-
-const appName = (remote ? remote.app : app || {}).name;
-const detail = multiline`
-  | Created by Jhen-Jie Hong
-  | (https://github.com/jhen0409)
-
-  | This software includes the following projects:
-
-  | https://github.com/facebook/react-devtools
-  | https://github.com/zalmoxisus/remotedev-app
-`;
-
-export const showAboutDialog = iconPath =>
-  (remote ? remote.dialog : dialog).showMessageBoxSync({
-    title: 'About',
-    message: `${appName} ${app.getVersion()}`,
-    detail,
-    icon: iconPath,
-    buttons: [],
-  });
-
 export const toggleDevTools = (win, type) => {
   if (!win || !type) return;
   if (type === 'chrome') {
@@ -48,7 +21,6 @@ export const resetZoom = win => {
     win.webContents.zoomLevel = 0;
   }
 };
-export const haveOpenedWindow = () => !!BrowserWindow.getAllWindows().length;
 export const toggleOpenInEditor = win =>
   win && win.webContents.executeJavaScript('window.toggleOpenInEditor()');
